[ etc ] pwa 설치 및 파일명 변경
