+++ api/formPreviewPage/createWishesLink.ts
@@ -3,6 +3,13 @@ import { client } from '../common/axios';
 import { WishesDataType } from '@/types/wishesDataType';
 
 export async function createWishesLink(wishesData: WishesDataType) {
+  const accessToken =
+    'eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiIxIiwiaWF0IjoxNjgzMTgxODk1LCJleHAiOjE2ODgwMjAyOTV9.RnUKwBtESVxvKU6YeXy4ssAJlEGHnIKd2LewByQhI8cPuTAh9tJQeJiMyRLYwJHFg8YwUZ667F1QRCECf1uR6w';
+
+  const startDate = wishesData.startDate.replaceAll('-', '.');
+  const endDate = wishesData.endDate.replaceAll('-', '.');
+  console.log(startDate, endDate);
+
   const data = await client.post(
     `${PATH.API}/${PATH.V1}/${PATH.WISHES}`,
     {
@@ -11,15 +18,18 @@ export async function createWishesLink(wishesData: WishesDataType) {
       title: wishesData.title,
       hint1: wishesData.hint1,
       hint2: wishesData.hint2,
-      startDate: wishesData.startDate,
-      endDate: wishesData.endDate,
+      startDate: startDate,
+      endDate: endDate,
       name: wishesData.name,
       bankName: wishesData.bankName,
       account: wishesData.account,
       phone: wishesData.phone,
     },
     {
-      headers: {},
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${accessToken}`,
+      },
     },
   );
   return data;
