+++ package.json
@@ -18,6 +18,7 @@
     "eslint": "^8.38.0",
     "eslint-config-next": "13.3.0",
     "next": "13.3.0",
+    "next-pwa": "^5.6.0",
     "postcss": "8.4.22",
     "react": "18.2.0",
     "react-datepicker": "^4.11.0",
