+++ pages/index.tsx
@@ -1,113 +1,85 @@
 import theme from '@/styles/theme';
 import styled from 'styled-components';
 import Image from 'next/image';
-import { SideBarIc } from '@/public/assets/icons';
-import { PillCakeImg } from '@/public/assets/images';
-import { MainChatImg } from '@/public/assets/images';
-import { GaugeBarImg } from '@/public/assets/images';
-import IconButton from '@/components/button/iconButton';
-import router from 'next/router';
-import ButtonBox from '@/components/button/buttonBox';
-import { useAuthKaKao } from '@/utils/hooks/useAuthKakao';
+import { GuideBtnIC, KakaoLoginIc } from '@/public/assets/icons';
+import { LoginCakeImg, LoginChatImg } from '@/public/assets/images';
+import Header from '@/components/common/header';
+import IconButton from '@/components/common/button/iconButton';
+import { useState } from 'react';
+import GuideModal from '@/components/modal/GuideModal';
 
-export default function MainPage() {
-  const moveToForm = () => {
-    router.push('/formPage');
-  };
+export default function LoginPage() {
+    const [showModal, setShowModal] = useState(false);
 
-  const { kakaoId, nickname, profileImage, email } = useAuthKaKao();
+    const clickModal = () => {
+        setShowModal(!showModal);
+    };
 
-  return (
-    <>
-      <Styled.HeaderContainer>
-        <Styled.Title>
-          {nickname}님,
-          <br />
-          <span style={{ color: theme.colors.main_blue }}>소원 링크</span>를 생성하고
-          <br />
-          케이크를 모아봐요!
-        </Styled.Title>
-        <Styled.SideContainer>
-          <IconButton src={SideBarIc} alt="사이드바" />
-          <Styled.DDay>D-?</Styled.DDay>
-        </Styled.SideContainer>
-      </Styled.HeaderContainer>
-      <h2>사용자 정보</h2>
-      <p>아이디: {kakaoId}</p>
-      <p>닉네임: {nickname}</p>
-      <p>프로필 이미지: {profileImage}</p>
-      <p>이메일: {email}</p>
+    const KakaoLogin = () => {
+        window.Kakao.Auth.authorize({
+            redirectUri: process.env.NEXT_PUBLIC_KAKAO_REDIRECT_URI,
+        });
+    };
 
-      <Styled.CenterContainer>
-        <Styled.BarContainer>
-          <Image src={GaugeBarImg} alt="게이지바" />
-        </Styled.BarContainer>
-        <Image src={MainChatImg} alt="날 모아줄래? (두근)" />
-        <Styled.ImageContainer>
-          <Image src={PillCakeImg} alt="케이크" />
-        </Styled.ImageContainer>
-        <Styled.About>모인 케이크 금액</Styled.About>
-        <Styled.AboutSmall>총 ???원</Styled.AboutSmall>
-      </Styled.CenterContainer>
+    return (
+        <>
+            <Header>
+                <IconButton onClick={clickModal} src={GuideBtnIC} alt="서비스 가이드" />
+            </Header>
+            {showModal && <GuideModal clickModal={clickModal} />}
 
-      <ButtonBox
-        handleClick={moveToForm}
-        backgroundColor={theme.colors.main_blue}
-        fontColor={theme.colors.white}
-      >
-        소원 링크 생성하기
-      </ButtonBox>
-    </>
-  );
+            <Styled.Container>
+                <Styled.ImageContainer>
+                    <Styled.Title>
+                        조물주보다
+                        <br />
+                        생일선물주
+                    </Styled.Title>
+
+                    <Image src={LoginChatImg} alt="진짜 원하는 선물을 말해봐요" />
+                    <Image src={LoginCakeImg} alt="로그인 케이크 이미지" />
+                </Styled.ImageContainer>
+                <Styled.About>사실 내가 갖고 싶었던 건...</Styled.About>
+                <Styled.AboutSmall>에X팟 맥스</Styled.AboutSmall>
+            </Styled.Container>
+
+            <IconButton onClick={KakaoLogin} src={KakaoLoginIc} alt="카카오 로그인" />
+        </>
+    );
 }
 
 const Styled = {
-  HeaderContainer: styled.div`
-    display: flex;
-  `,
-
-  SideContainer: styled.div`
-    margin-left: auto;
-  `,
-
-  Title: styled.div`
-    ${theme.fonts.headline24_130};
-    margin: 0 0 3rem;
-  `,
-
-  DDay: styled.div`
-    ${theme.fonts.headline20};
+    Title: styled.div`
+    ${theme.fonts.title56};
     color: ${theme.colors.main_blue};
+    margin: 2rem 0 2.8rem;
     display: flex;
     justify-content: center;
-    margin-top: 2.3rem;
+    white-space: pre-line;
   `,
 
-  CenterContainer: styled.div`
-    margin: 9rem 0 15.5rem;
-  `,
+    Container: styled.div``,
 
-  BarContainer: styled.div`
-    float: right;
-    margin: 6rem 1.5rem 0 0;
-  `,
-
-  ImageContainer: styled.div`
+    ImageContainer: styled.div`
     width: 100%;
-    text-align: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
   `,
 
-  About: styled.div`
+    About: styled.div`
     ${theme.fonts.headline24_100};
     color: ${theme.colors.main_blue};
-    margin: 0 0 1rem;
+    margin: 1.4rem 0 1.5rem;
     display: flex;
     justify-content: center;
   `,
 
-  AboutSmall: styled.div`
-    ${theme.fonts.headline24_100};
+    AboutSmall: styled.div`
+    ${theme.fonts.body16};
+    color: ${theme.colors.main_blue};
     display: flex;
     justify-content: center;
+    margin: 0 0 10.15rem;
   `,
 };
